wip: refactoring, simplify, part 2
