+++ .github/scripts/junit/README.md
@@ -6,13 +6,14 @@ Tools for parsing JUnit XML test reports and calculating optimal test chunk dist
 
 This package provides:
 
-1. **`parse-test-timings.ts`** - Parse JUnit XML reports to extract timing data
-2. **`calculate-optimal-chunks.ts`** - Distribute tests across CI chunks using bin-packing
+1. **`src/parse-test-timings.ts`** - Parse JUnit XML reports to extract timing data
+2. **`src/calculate-optimal-chunks.ts`** - Distribute tests across CI chunks using bin-packing
+3. **`src/sanitize-junit-xml.ts`** - Strip unnecessary data from JUnit XML for efficient storage
 
 ## Installation
 
 ```bash
-	cd .github/scripts/junit
+cd .github/scripts/junit
 bun install
 ```
 
@@ -22,23 +23,23 @@ bun install
 
 ```bash
 # File-level timing (v1.0 format)
-bun parse-test-timings.ts .test-timings.json report/*.xml
+bun src/parse-test-timings.ts .test-timings.json report/*.xml
 
 # Example-level timing (v2.0 format)
-bun parse-test-timings.ts .test-timings.json report/*.xml --granularity=example
+bun src/parse-test-timings.ts .test-timings.json report/*.xml --granularity=example
 ```
 
 ### Calculate Optimal Chunks
 
 ```bash
 # Get spec files for chunk 0 of 4
-bun calculate-optimal-chunks.ts .test-timings.json 4 0
+bun src/calculate-optimal-chunks.ts .test-timings.json 4 0
 
 # With example-level granularity
-bun calculate-optimal-chunks.ts .test-timings.json 4 0 --granularity=example
+bun src/calculate-optimal-chunks.ts .test-timings.json 4 0 --granularity=example
 
 # Hybrid mode (auto-split large files)
-bun calculate-optimal-chunks.ts .test-timings.json 4 0 --granularity=hybrid
+bun src/calculate-optimal-chunks.ts .test-timings.json 4 0 --granularity=hybrid
 ```
 
 ## Testing
@@ -73,25 +74,34 @@ bun run typecheck
 .github/scripts/junit/
 ├── package.json              # Package configuration
 ├── tsconfig.json             # TypeScript configuration
-├── parse-test-timings.ts     # Main timing parser script
-├── calculate-optimal-chunks.ts  # Main chunk calculator script
-├── lib/
-│   ├── index.ts              # Library exports
-│   ├── parser.ts             # JUnit XML parsing functions
-│   └── chunker.ts            # Bin-packing algorithm and utilities
-├── __tests__/
-│   ├── parser.test.ts        # Parser unit tests
-│   ├── chunker.test.ts       # Chunker unit tests
-│   └── integration.test.ts   # End-to-end integration tests
-└── __fixtures__/
-    ├── sample-results.xml    # Sample JUnit XML for testing
-    ├── sample-results-2.xml  # Additional sample XML
-    └── empty-results.xml     # Edge case: empty results
+└── src/
+    ├── index.ts              # Library exports
+    ├── parser.ts             # JUnit XML parsing functions
+    ├── parser.test.ts        # Parser tests (unit + e2e)
+    ├── chunker.ts            # Bin-packing algorithm and utilities
+    ├── chunker.test.ts       # Chunker tests (unit + e2e)
+    ├── parse-test-timings.ts               # Main timing parser script
+    ├── parse-test-timings.test.ts          # Parser script tests (e2e)
+    ├── calculate-optimal-chunks.ts         # Main chunk calculator script
+    ├── calculate-optimal-chunks.test.ts    # Chunk calculator tests (e2e)
+    ├── sanitize-junit-xml.ts               # XML sanitizer script
+    ├── sanitize-junit-xml.test.ts          # Sanitizer tests (e2e)
+    └── __fixtures__/
+        ├── sample-results.xml    # Sample JUnit XML for testing
+        ├── sample-results-2.xml  # Additional sample XML
+        └── empty-results.xml     # Edge case: empty results
 ```
+
+### Test Naming Convention
+
+Tests follow the `{source_file}.test.ts` pattern:
+- Each source file has a corresponding test file
+- Unit tests are in their own `describe()` blocks
+- End-to-end tests are in `describe("e2e", ...)` blocks
 
 ## Library API
 
-The core functions are exported from `lib/` for use in other scripts:
+The core functions are exported from `src/index.ts` for use in other scripts:
 
 ```typescript
 import {
@@ -107,7 +117,7 @@ import {
   buildFileItemsFromTimings,
   collapseExampleOutput,
   parseChunkArgs,
-} from "./lib";
+} from "./src/index";
 ```
 
 ## Timing Data Formats
