--- .github/scripts/junit/__tests__/chunk-tests-baseline-source.test.ts
@@ -1,83 +1,0 @@
-/**
- * Verify chunk-tests.sh prefers a versioned baseline timing file when present.
- */
-
-import { describe, test, expect } from "bun:test";
-import { existsSync, readFileSync, rmSync, mkdirSync, writeFileSync, readdirSync } from "node:fs";
-import { join, relative } from "node:path";
-import { spawnSync } from "node:child_process";
-
-function listSpecFiles(projectRoot: string): string[] {
-    const specRoot = join(projectRoot, "spec");
-    const results: string[] = [];
-
-    function walk(dir: string) {
-        for (const entry of readdirSync(dir, { withFileTypes: true })) {
-            const fullPath = join(dir, entry.name);
-            if (entry.isDirectory()) {
-                walk(fullPath);
-                continue;
-            }
-            if (entry.isFile() && entry.name.endsWith("_spec.sh")) {
-                const rel = relative(projectRoot, fullPath).replaceAll("\\", "/");
-                results.push(rel);
-            }
-        }
-    }
-
-    walk(specRoot);
-    return results.sort();
-}
-
-describe(".github/scripts/chunk-tests.sh (baseline timing file selection)", () => {
-    test("uses .github/data/linux/test-timings.json when RUNNER_OS=Linux and file exists", () => {
-        const projectRoot = join(import.meta.dir, "../../../..");
-        const specFiles = listSpecFiles(projectRoot);
-        expect(specFiles.length).toBeGreaterThan(0);
-
-        const baselineDir = join(projectRoot, ".github/data/linux");
-        const baselineTimingFile = join(baselineDir, "test-timings.json");
-
-        const hadExistingBaseline = existsSync(baselineTimingFile);
-        const baselineBackup = hadExistingBaseline ? readFileSync(baselineTimingFile, "utf-8") : null;
-
-        try {
-            mkdirSync(baselineDir, { recursive: true });
-
-            const timings: Record<string, number> = {};
-            for (const file of specFiles) timings[file] = 1.0;
-
-            writeFileSync(
-                baselineTimingFile,
-                JSON.stringify(
-                    {
-                        version: "1.0",
-                        description: "Baseline timings (generated by tests)",
-                        timings,
-                        total_time: specFiles.length,
-                        file_count: specFiles.length,
-                        source_files: ["generated"],
-                    },
-                    null,
-                    2
-                )
-            );
-
-            const result = spawnSync("bash", [".github/scripts/chunk-tests.sh", "4", "0"], {
-                cwd: projectRoot,
-                encoding: "utf-8",
-                env: { ...process.env, RUNNER_OS: "Linux" },
-            });
-
-            expect(result.status).toBe(0);
-            expect(result.stderr).toContain("📈 Timing source:");
-            expect(result.stderr).toContain(".github/data/linux/test-timings.json");
-        } finally {
-            if (hadExistingBaseline && baselineBackup !== null) {
-                writeFileSync(baselineTimingFile, baselineBackup);
-            } else if (existsSync(baselineTimingFile)) {
-                rmSync(baselineTimingFile, { force: true });
-            }
-        }
-    });
-});
