--- .github/scripts/junit/__tests__/chunk-tests-estimates.test.ts
@@ -1,82 +1,0 @@
-/**
- * Integration-style test ensuring CI chunking prints time estimates before tests run.
- */
-
-import { describe, test, expect } from "bun:test";
-import { existsSync, readFileSync, rmSync, writeFileSync, readdirSync } from "node:fs";
-import { join, relative } from "node:path";
-import { spawnSync } from "node:child_process";
-
-function listSpecFiles(projectRoot: string): string[] {
-    const specRoot = join(projectRoot, "spec");
-    const results: string[] = [];
-
-    function walk(dir: string) {
-        for (const entry of readdirSync(dir, { withFileTypes: true })) {
-            const fullPath = join(dir, entry.name);
-            if (entry.isDirectory()) {
-                walk(fullPath);
-                continue;
-            }
-            if (entry.isFile() && entry.name.endsWith("_spec.sh")) {
-                const rel = relative(projectRoot, fullPath).replaceAll("\\", "/");
-                results.push(rel);
-            }
-        }
-    }
-
-    walk(specRoot);
-    return results.sort();
-}
-
-describe(".github/scripts/chunk-tests.sh (timing visualization)", () => {
-    test("prints per-chunk estimate and full distribution table to stderr", () => {
-        const projectRoot = join(import.meta.dir, "../../../..");
-        const timingFile = join(projectRoot, ".test-timings.json");
-
-        const hadExistingTimingFile = existsSync(timingFile);
-        const timingFileBackup = hadExistingTimingFile ? readFileSync(timingFile, "utf-8") : null;
-
-        try {
-            const specFiles = listSpecFiles(projectRoot);
-            expect(specFiles.length).toBeGreaterThan(0);
-
-            const timings: Record<string, number> = {};
-            for (const file of specFiles) {
-                timings[file] = 1.0;
-            }
-
-            writeFileSync(
-                timingFile,
-                JSON.stringify(
-                    {
-                        version: "1.0",
-                        description: "Test timings (generated by tests)",
-                        timings,
-                        total_time: specFiles.length,
-                        file_count: specFiles.length,
-                        source_files: ["generated"],
-                    },
-                    null,
-                    2
-                )
-            );
-
-            const result = spawnSync(
-                "bash",
-                [".github/scripts/chunk-tests.sh", "4", "0", "--granularity=file"],
-                { cwd: projectRoot, encoding: "utf-8" }
-            );
-
-            expect(result.status).toBe(0);
-            expect(result.stderr).toContain("⏱️");
-            expect(result.stderr).toContain("📊 Chunk distribution");
-        } finally {
-            if (hadExistingTimingFile && timingFileBackup !== null) {
-                writeFileSync(timingFile, timingFileBackup);
-            } else if (existsSync(timingFile)) {
-                rmSync(timingFile, { force: true });
-            }
-        }
-    });
-});
