+++ .github/scripts/junit/src/__fixtures__/empty-results.xml
