+++ .github/scripts/junit/src/sanitize-junit-xml.ts
