+++ .github/workflows/baseline.yaml
@@ -107,8 +107,8 @@ jobs:
             echo "No JUnit XML files found in report/; cannot generate baseline."
             exit 1
           fi
-          bun .github/scripts/junit/sanitize-junit-xml.ts .github/data/linux/report/baseline.xml "${XML_FILES[@]}"
-          bun .github/scripts/junit/parse-test-timings.ts .github/data/linux/test-timings.json .github/data/linux/report/baseline.xml --granularity=example
+          bun .github/scripts/junit/src/sanitize-junit-xml.ts .github/data/linux/report/baseline.xml "${XML_FILES[@]}"
+          bun .github/scripts/junit/src/parse-test-timings.ts .github/data/linux/test-timings.json .github/data/linux/report/baseline.xml --granularity=example
 
       - name: Upload baseline artifact (Linux)
         uses: actions/upload-artifact@v4
@@ -179,8 +179,8 @@ jobs:
             echo "No JUnit XML files found in report/; cannot generate baseline."
             exit 1
           fi
-          bun .github/scripts/junit/sanitize-junit-xml.ts .github/data/macos/report/baseline.xml "${XML_FILES[@]}"
-          bun .github/scripts/junit/parse-test-timings.ts .github/data/macos/test-timings.json .github/data/macos/report/baseline.xml --granularity=example
+          bun .github/scripts/junit/src/sanitize-junit-xml.ts .github/data/macos/report/baseline.xml "${XML_FILES[@]}"
+          bun .github/scripts/junit/src/parse-test-timings.ts .github/data/macos/test-timings.json .github/data/macos/report/baseline.xml --granularity=example
 
       - name: Upload baseline artifact (macOS)
         uses: actions/upload-artifact@v4
